requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
pgvector>=0.3.0
openai>=1.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
cryptography>=42.0.5
pydantic-settings>=2.2.1
beautifulsoup4>=4.12.0
numpy>=1.26.0
